// types/codegenTypes.ts
export type Maybe<T> = T | null
// ...
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  Date: any
  ObjectID: any
  timestamptz: any
  uuid: any
}
// ...
export type Launch = {
  __typename?: 'Launch'
  details: Maybe<Scalars['String']>
  id: Maybe<Scalars['ID']>
  is_tentative: Maybe<Scalars['Boolean']>
  launch_date_local: Maybe<Scalars['Date']>
  launch_date_unix: Maybe<Scalars['Date']>
  launch_date_utc: Maybe<Scalars['Date']>
  launch_site: Maybe<LaunchSite>
  launch_success: Maybe<Scalars['Boolean']>
  launch_year: Maybe<Scalars['String']>
  links: Maybe<LaunchLinks>
  mission_id: Maybe<Array<Maybe<Scalars['String']>>>
  mission_name: Maybe<Scalars['String']>
  rocket: Maybe<LaunchRocket>
  static_fire_date_unix: Maybe<Scalars['Date']>
  static_fire_date_utc: Maybe<Scalars['Date']>
  telemetry: Maybe<LaunchTelemetry>
  tentative_max_precision: Maybe<Scalars['String']>
  upcoming: Maybe<Scalars['Boolean']>
  ships: Maybe<Array<Maybe<Ship>>>
}
// ...
export type LaunchLinks = {
  __typename?: 'LaunchLinks'
  article_link: Maybe<Scalars['String']>
  flickr_images: Maybe<Array<Maybe<Scalars['String']>>>
  mission_patch_small: Maybe<Scalars['String']>
  mission_patch: Maybe<Scalars['String']>
  presskit: Maybe<Scalars['String']>
  reddit_campaign: Maybe<Scalars['String']>
  reddit_launch: Maybe<Scalars['String']>
  reddit_media: Maybe<Scalars['String']>
  reddit_recovery: Maybe<Scalars['String']>
  video_link: Maybe<Scalars['String']>
  wikipedia: Maybe<Scalars['String']>
}

export type LaunchRocket = {
  __typename?: 'LaunchRocket'
  fairings: Maybe<LaunchRocketFairings>
  first_stage: Maybe<LaunchRocketFirstStage>
  rocket_name: Maybe<Scalars['String']>
  rocket_type: Maybe<Scalars['String']>
  rocket: Maybe<Rocket>
  second_stage: Maybe<LaunchRocketSecondStage>
}

export type LaunchRocketFairings = {
  __typename?: 'LaunchRocketFairings'
  recovered: Maybe<Scalars['Boolean']>
  recovery_attempt: Maybe<Scalars['Boolean']>
  reused: Maybe<Scalars['Boolean']>
  ship: Maybe<Scalars['String']>
}

export type LaunchRocketFirstStage = {
  __typename?: 'LaunchRocketFirstStage'
  cores: Maybe<Array<Maybe<LaunchRocketFirstStageCore>>>
}
// ...
export type LaunchRocketSecondStage = {
  __typename?: 'LaunchRocketSecondStage'
  block: Maybe<Scalars['Int']>
  payloads: Maybe<Array<Maybe<Payload>>>
}

export type LaunchSite = {
  __typename?: 'LaunchSite'
  site_id: Maybe<Scalars['String']>
  site_name_long: Maybe<Scalars['String']>
  site_name: Maybe<Scalars['String']>
}

export type LaunchTelemetry = {
  __typename?: 'LaunchTelemetry'
  flight_club: Maybe<Scalars['String']>
}
// ...
export type Rocket = {
  __typename?: 'Rocket'
  active: Maybe<Scalars['Boolean']>
  boosters: Maybe<Scalars['Int']>
  company: Maybe<Scalars['String']>
  cost_per_launch: Maybe<Scalars['Int']>
  country: Maybe<Scalars['String']>
  description: Maybe<Scalars['String']>
  diameter: Maybe<Distance>
  engines: Maybe<RocketEngines>
  first_flight: Maybe<Scalars['Date']>
  first_stage: Maybe<RocketFirstStage>
  height: Maybe<Distance>
  id: Maybe<Scalars['ID']>
  landing_legs: Maybe<RocketLandingLegs>
  mass: Maybe<Mass>
  name: Maybe<Scalars['String']>
  payload_weights: Maybe<Array<Maybe<RocketPayloadWeight>>>
  second_stage: Maybe<RocketSecondStage>
  stages: Maybe<Scalars['Int']>
  success_rate_pct: Maybe<Scalars['Int']>
  type: Maybe<Scalars['String']>
  wikipedia: Maybe<Scalars['String']>
}
// ...
export type Ship = {
  __typename?: 'Ship'
  abs: Maybe<Scalars['Int']>
  active: Maybe<Scalars['Boolean']>
  attempted_landings: Maybe<Scalars['Int']>
  class: Maybe<Scalars['Int']>
  course_deg: Maybe<Scalars['Int']>
  home_port: Maybe<Scalars['String']>
  id: Maybe<Scalars['ID']>
  image: Maybe<Scalars['String']>
  imo: Maybe<Scalars['Int']>
  missions: Maybe<Array<Maybe<ShipMission>>>
  mmsi: Maybe<Scalars['Int']>
  model: Maybe<Scalars['String']>
  name: Maybe<Scalars['String']>
  position: Maybe<ShipLocation>
  roles: Maybe<Array<Maybe<Scalars['String']>>>
  speed_kn: Maybe<Scalars['Float']>
  status: Maybe<Scalars['String']>
  successful_landings: Maybe<Scalars['Int']>
  type: Maybe<Scalars['String']>
  url: Maybe<Scalars['String']>
  weight_kg: Maybe<Scalars['Int']>
  weight_lbs: Maybe<Scalars['Int']>
  year_built: Maybe<Scalars['Int']>
}
// ...
export type LaunchQuery = {
  launch: Maybe<
    { __typename?: 'Launch' } & Pick<Launch, 'id' | 'mission_name' | 'launch_date_local'> & {
        rocket: Maybe<{ __typename?: 'LaunchRocket' } & Pick<LaunchRocket, 'rocket_name'>>
        links: Maybe<{ __typename?: 'LaunchLinks' } & Pick<LaunchLinks, 'video_link'>>
      }
  >
}